% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch-gsub.R
\name{batch_gsub}
\alias{batch_gsub}
\title{Batch File Text Replacement}
\usage{
batch_gsub(pattern, replacement, path, regexp = "[.](R|r)$",
  ignore.case = FALSE, fixed = FALSE, recurse = FALSE,
  progress = FALSE, report = TRUE, options = furrr::future_options(),
  ask = getOption("batchr.ask", TRUE))
}
\arguments{
\item{pattern}{A string of the regular expression to match.}

\item{replacement}{A string of the replacement text.}

\item{path}{A string of the path to the directory with the files for processing.}

\item{regexp}{A string of a regular expression. Only non-hidden file names
which match the regular expression will be batch processed.}

\item{ignore.case}{A flag specifying whether to ignore the case.}

\item{fixed}{A flag specifying whether to match \code{pattern} as is
(as opposed to as a regular expression).}

\item{recurse}{A flag specifying whether to recurse into path's subdirectories.}

\item{progress}{A flag specifying whether to print a progress bar.}

\item{report}{A flag specifying whether to outputs a report of the status of individual files to the console.}

\item{options}{The future specific options to use with the workers.
seed must be \code{FALSE}.}

\item{ask}{A flag specifying whether to ask before starting to process the files.}
}
\value{
An invisible flag indicating whether all the files were
successfully processed.
}
\description{
Uses \code{\link[batchr:batch_process]{batchr::batch_process()}} and \code{\link[=gsub_file]{gsub_file()}} to
perform batch text file replacement.
By default it replaces text in all .R and .r files in the working directory.
}
\seealso{
\code{\link[batchr:batch_process]{batchr::batch_process()}} and \code{\link[=gsub_file]{gsub_file()}}
}
