% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchgsub-package.R
\docType{package}
\name{batchgsub-package}
\alias{batchgsub}
\alias{batchgsub-package}
\title{batchgsub: Batch File Text Replacement}
\description{
A Shiny app and Rstudio addin to batch process text replacement in content of files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/batchgsub}
  \item Report bugs at \url{https://github.com/poissonconsulting/batchgsub/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID})

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
