% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsub-file.R
\name{gsub_file}
\alias{gsub_file}
\title{File Text Replacement}
\usage{
gsub_file(file, pattern, replacement, ignore.case = FALSE, fixed = FALSE)
}
\arguments{
\item{file}{A string of the name of the file to modify.}

\item{pattern}{A string of the regular expression to match.}

\item{replacement}{A string of the replacement text.}

\item{ignore.case}{A flag specifying whether to ignore the case.}

\item{fixed}{A flag specifying whether to match \code{pattern} as is
(as opposed to as a regular expression).}
}
\value{
TRUE
}
\description{
Uses \code{\link[base:gsub]{base::gsub()}} to perform text replacement on a file.
}
\seealso{
\code{\link[base:gsub]{base::gsub()}}
}
