% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the batch_gsub Shiny Application}
\usage{
run_app(path = getwd())
}
\arguments{
\item{path}{A character string of the default path.}
}
\description{
Run the batch_gsub Shiny Application
}
